Don't use latest.
Latest was a mistake in Docker. Folks think that it moves but it
doesn't. They'll use it anyway. So we will set it to be identical to
main (it will move with every commit).

The releases however will only be variants of major.minor.patch. This
will allow folks to pin to major or major.minor and go along for the
ride. As those tags will move. For folks that want to pin to a specific
release, they can use major.minor.patch.

Signed-off-by: Chad Metcalf <metcalfc@gmail.com>